fix(support): remove PostHog chat overlay
Summary:
- remove PostHog Conversations imports and open handlers from Web and
  Electron clients
- route Support actions to Matrix documentation while retaining bounded
  PostHog analytics
- disable Conversations explicitly and add cross-surface regression tests

Rationale:
- the provider-owned support panel is an unintended OS-level overlay
- an explicit disable flag prevents project remote config from restoring it

Tests:
- focused Vitest suite: 38 passed
- shell and Desktop renderer TypeScript checks
- bun run check:patterns (0 violations)
