+++ desktop/src/renderer/src/features/support/DesktopSupportButton.tsx
@@ -1,9 +1,10 @@
-import { MessageCircle } from "../../lib/hugeicons";
-import {
-  openDesktopSupport,
-} from "./DesktopSupportWidget";
+import { CircleHelp } from "../../lib/hugeicons";
+import { useTabs } from "../../stores/tabs";
+import { openHelpInMatrixBrowser } from "../browser/help-navigation";
 
 export default function DesktopSupportButton() {
+  const openTab = useTabs((state) => state.openTab);
+
   return (
     <button
       type="button"
@@ -12,10 +13,10 @@ export default function DesktopSupportButton() {
       className="flex size-7 shrink-0 items-center justify-center rounded-md outline-none transition-colors hover:bg-[var(--bg-hover)] focus-visible:bg-[var(--bg-hover)]"
       style={{ color: "var(--text-secondary)" }}
       onClick={() => {
-        void openDesktopSupport();
+        openHelpInMatrixBrowser(openTab);
       }}
     >
-      <MessageCircle aria-hidden="true" size={16} />
+      <CircleHelp aria-hidden="true" size={16} />
     </button>
   );
 }
