+++ desktop/src/renderer/src/vite-env.d.ts
@@ -1,7 +1,5 @@
 /// <reference types="vite/client" />
 
-declare module "posthog-js/dist/conversations";
-
 declare module "*?worker" {
   const workerConstructor: new () => Worker;
   export default workerConstructor;
