+++ shell/src/components/desktop/WebDesktopControls.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Link from "next/link";
-import { MessageCircleIcon, SearchIcon, ServerIcon } from "@/lib/hugeicons";
+import { CircleHelpIcon, SearchIcon, ServerIcon } from "@/lib/hugeicons";
 import { UserButton } from "../UserButton";
 
 export type WebDesktopSettingsSection = "appearance" | "billing" | "integrations";
@@ -9,14 +9,13 @@ export type WebDesktopSettingsSection = "appearance" | "billing" | "integrations
 interface WebDesktopControlsProps {
   onOpenSettings: (section: WebDesktopSettingsSection) => void;
   onOpenCommandPalette: () => void;
-  onOpenSupport: () => void;
 }
 
 const actionClass =
   "flex size-7 items-center justify-center rounded-full border border-border/60 bg-card text-muted-foreground shadow-sm transition-colors hover:bg-muted/70 hover:text-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring";
 
 /** Web-safe equivalents of the native Desktop titlebar controls. */
-export function WebDesktopControls({ onOpenSettings, onOpenCommandPalette, onOpenSupport }: WebDesktopControlsProps) {
+export function WebDesktopControls({ onOpenSettings, onOpenCommandPalette }: WebDesktopControlsProps) {
   return (
     <nav aria-label="Desktop controls" className="flex items-center gap-1.5">
       <button
@@ -28,15 +27,16 @@ export function WebDesktopControls({ onOpenSettings, onOpenCommandPalette, onOpe
       >
         <SearchIcon className="size-3.5" aria-hidden="true" />
       </button>
-      <button
-        type="button"
-        aria-label="Support chat"
-        title="Support chat"
+      <a
+        href="https://matrix-os.com/docs"
+        target="_blank"
+        rel="noreferrer"
+        aria-label="Support"
+        title="Support"
         className={actionClass}
-        onClick={onOpenSupport}
       >
-        <MessageCircleIcon className="size-3.5" aria-hidden="true" />
-      </button>
+        <CircleHelpIcon className="size-3.5" aria-hidden="true" />
+      </a>
       <Link
         href="/runtime"
         aria-label="Switch computer"
