+++ shell/src/lib/posthog-client.ts
@@ -1,6 +1,5 @@
 "use client";
 
-import "posthog-js/dist/conversations";
 import posthog from "posthog-js";
 import {
   buildPostHogCookieConsentInitOptions,
@@ -23,10 +22,6 @@ const config = getPostHogClientConfig({
   NEXT_PUBLIC_POSTHOG_API_HOST: process.env.NEXT_PUBLIC_POSTHOG_API_HOST ?? "/relay",
 });
 const CLIENT_ERROR_REPORT_TIMEOUT_MS = 10_000;
-const SUPPORT_AVAILABILITY_TIMEOUT_MS = 10_000;
-const POSTHOG_WIDGET_ID = "ph-conversations-widget-container";
-const POSTHOG_LAUNCHER_SELECTOR = 'button[aria-label^="Open chat"]';
-const POSTHOG_CLOSE_SELECTOR = 'button[aria-label="Close"]';
 // Replay kill switch. NEXT_PUBLIC_* is inlined at build time, so the build
 // flag alone cannot stop replay during an incident. The layout additionally
 // exposes the server's runtime POSTHOG_DISABLE_REPLAY env as a data
@@ -91,47 +86,6 @@ export function capturePostHogEvent(event: string, properties: ClientProperties 
   } catch (err: unknown) {
     console.warn("[posthog] Failed to capture client event:", err instanceof Error ? err.name : typeof err);
   }
-}
-
-export async function openShellSupport(currentConfig: typeof config = config): Promise<boolean> {
-  if (!currentConfig) return false;
-  try {
-    ensurePostHogInitialized(currentConfig);
-    const deadline = Date.now() + SUPPORT_AVAILABILITY_TIMEOUT_MS;
-    while (Date.now() < deadline) {
-      if (posthog.conversations.isAvailable()) {
-        posthog.conversations.show();
-        let closeButton = findPostHogButton(POSTHOG_CLOSE_SELECTOR);
-        if (!closeButton) {
-          const launcher = await waitForPostHogButton(POSTHOG_LAUNCHER_SELECTOR, deadline);
-          if (!launcher) return false;
-          launcher.click();
-          closeButton = await waitForPostHogButton(POSTHOG_CLOSE_SELECTOR, deadline);
-        }
-        if (!closeButton) return false;
-        posthog.capture("shell_support_chat_opened", { matrix_client: "web" });
-        return true;
-      }
-      await new Promise((resolve) => globalThis.setTimeout(resolve, 100));
-    }
-    return false;
-  } catch (err: unknown) {
-    console.warn("[posthog] Failed to open support chat:", err instanceof Error ? err.name : typeof err);
-    return false;
-  }
-}
-
-async function waitForPostHogButton(selector: string, deadline: number): Promise<HTMLButtonElement | null> {
-  while (Date.now() < deadline) {
-    const button = findPostHogButton(selector);
-    if (button) return button;
-    await new Promise((resolve) => globalThis.setTimeout(resolve, 50));
-  }
-  return null;
-}
-
-function findPostHogButton(selector: string): HTMLButtonElement | null {
-  return document.getElementById(POSTHOG_WIDGET_ID)?.querySelector<HTMLButtonElement>(selector) ?? null;
 }
 
 export function setPostHogPersonPropertiesOnce(
@@ -222,6 +176,7 @@ export function initializeShellPostHog(
     capture_exceptions: true,
     capture_dead_clicks: false,
     rageclick: false,
+    disable_conversations: true,
     // Masked session replay: all inputs masked, opt-in text masking via
     // [data-ph-mask], and sensitive surfaces (terminal, chat transcripts,
     // file listings) blocked entirely via the ph-no-capture class, which
