+++ tests/shell/posthog-proxy.test.ts
@@ -6,10 +6,6 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import nextConfig from "../../shell/next.config";
 
 const posthogMock = vi.hoisted(() => ({
-  conversations: {
-    isAvailable: vi.fn(() => true),
-    show: vi.fn(),
-  },
   init: vi.fn(),
   capture: vi.fn(),
   identify: vi.fn(),
@@ -20,7 +16,6 @@ const posthogMock = vi.hoisted(() => ({
 vi.mock("posthog-js", () => ({
   default: posthogMock,
 }));
-vi.mock("posthog-js/dist/conversations", () => ({}));
 
 type RewriteRule = { source: string; destination: string };
 
@@ -32,9 +27,7 @@ async function getRewrites(): Promise<RewriteRule[]> {
 
 describe("shell PostHog same-origin proxy", () => {
   afterEach(() => {
-    document.getElementById("ph-conversations-widget-container")?.remove();
-    document.getElementById("unrelated-close")?.remove();
-    posthogMock.conversations.show.mockReset();
+    vi.clearAllMocks();
   });
 
   it("rewrites the same-origin health probe to the gateway", async () => {
@@ -86,6 +79,7 @@ describe("shell PostHog same-origin proxy", () => {
     expect(token).toBe("phc_test");
     expect(options.api_host).toBe("/relay");
     expect(options.ui_host).toBe("https://eu.posthog.com");
+    expect(options.disable_conversations).toBe(true);
   });
 
   it("only resets identity for provably identified sessions", async () => {
@@ -112,41 +106,12 @@ describe("shell PostHog same-origin proxy", () => {
     expect(posthogMock.reset).toHaveBeenCalledTimes(1);
   });
 
-  it("opens PostHog Conversations from the shell navbar", async () => {
-    const unrelatedClose = document.createElement("button");
-    unrelatedClose.id = "unrelated-close";
-    unrelatedClose.setAttribute("aria-label", "Close");
-    document.body.appendChild(unrelatedClose);
-    const launcherClick = vi.fn(() => {
-      const close = document.createElement("button");
-      close.setAttribute("aria-label", "Close");
-      document.getElementById("ph-conversations-widget-container")?.replaceChildren(close);
-    });
-    posthogMock.conversations.show.mockImplementation(() => {
-      const container = document.createElement("div");
-      container.id = "ph-conversations-widget-container";
-      const launcher = document.createElement("button");
-      launcher.setAttribute("aria-label", "Open chat");
-      launcher.addEventListener("click", launcherClick);
-      container.appendChild(launcher);
-      document.body.appendChild(container);
-    });
-    const { openShellSupport } = await import("../../shell/src/lib/posthog-client");
-    const opened = await openShellSupport({
-      token: "phc_test",
-      apiHost: "/relay",
-      uiHost: "https://eu.posthog.com",
-    });
+  it("does not bundle or expose the PostHog Conversations overlay", () => {
+    const source = readPostHogClientSource();
 
-    expect(opened).toBe(true);
-    expect(posthogMock.conversations.show).toHaveBeenCalledOnce();
-    expect(launcherClick).toHaveBeenCalledOnce();
-    expect(
-      document.querySelector('#ph-conversations-widget-container button[aria-label="Close"]'),
-    ).not.toBeNull();
-    expect(posthogMock.capture).toHaveBeenCalledWith("shell_support_chat_opened", {
-      matrix_client: "web",
-    });
+    expect(source).not.toContain("posthog-js/dist/conversations");
+    expect(source).not.toContain("openShellSupport");
+    expect(source).not.toContain("posthog.conversations");
   });
 
   it("defaults the shell api host to the /relay same-origin proxy", () => {
