+++ tests/shell/web-desktop-controls.test.tsx
@@ -15,23 +15,21 @@ describe("WebDesktopControls", () => {
   it("maps native Desktop right-side actions to web-safe navigation", () => {
     const onOpenSettings = vi.fn();
     const onOpenCommandPalette = vi.fn();
-    const onOpenSupport = vi.fn();
     render(
       <WebDesktopControls
         onOpenSettings={onOpenSettings}
         onOpenCommandPalette={onOpenCommandPalette}
-        onOpenSupport={onOpenSupport}
       />,
     );
 
     fireEvent.click(screen.getByRole("button", { name: "Search" }));
     expect(onOpenCommandPalette).toHaveBeenCalledOnce();
     expect(screen.getByRole("link", { name: "Switch computer" }).getAttribute("href")).toBe("/runtime");
-    fireEvent.click(screen.getByRole("button", { name: "Support chat" }));
-    expect(onOpenSupport).toHaveBeenCalledOnce();
+    expect(screen.getByRole("link", { name: "Support" }).getAttribute("href"))
+      .toBe("https://matrix-os.com/docs");
     expect(Array.from(screen.getByRole("navigation", { name: "Desktop controls" }).children)
       .map((control) => control.getAttribute("aria-label") ?? control.textContent))
-      .toEqual(["Search", "Support chat", "Switch computer", "Account"]);
+      .toEqual(["Search", "Support", "Switch computer", "Account"]);
     fireEvent.click(screen.getByRole("button", { name: "Account" }));
     expect(onOpenSettings).toHaveBeenCalledWith("billing");
   });
